+++ src/Elcodi/Bundle/ProductBundle/Resources/config/classes.yml
@@ -17,7 +17,6 @@ parameters:
     #
     # Factories
     #
-    elcodi.core.product.factory.abstract.abstract.class: Elcodi\Component\Product\Factory\Abstracts\AbstractPurchasableFactory
     elcodi.core.product.factory.product.class: Elcodi\Component\Product\Factory\ProductFactory
     elcodi.core.product.factory.variant.class: Elcodi\Component\Product\Factory\VariantFactory
     elcodi.core.product.factory.manufacturer.class: Elcodi\Component\Product\Factory\ManufacturerFactory
