+++ src/Elcodi/Component/Product/Factory/ProductFactory.php
@@ -21,7 +21,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 
 use Elcodi\Component\Product\ElcodiProductTypes;
 use Elcodi\Component\Product\Entity\Product;
-use Elcodi\Component\Product\Factory\Abstracts\AbstractPurchasableFactory;
+use Elcodi\Component\Currency\Factory\Abstracts\AbstractPurchasableFactory;
 
 /**
  * Factory for Product entities
